requirements.txt
# Core Framework
streamlit==1.31.0
python-dotenv==1.0.1

# AI & ML
openai==1.12.0
langchain==0.1.6
langchain-openai==0.0.5
transformers==4.37.2
torch==2.2.0
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0

# Firebase
firebase-admin==6.4.0
pyrebase4==4.7.1

# Authentication & Security
PyJWT==2.8.0
cryptography==42.0.2
pycryptodome==3.20.0
bcrypt==4.1.2

# Database
google-cloud-firestore==2.14.0

# Data Visualization
matplotlib==3.8.2
plotly==5.18.0
seaborn==0.13.2

# HTTP & API
requests==2.31.0
httpx==0.26.0
aiohttp==3.9.3

# Monitoring & Logging
sentry-sdk==1.40.0
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-cov==4.1.0
pytest-asyncio==0.23.4
pytest-mock==3.12.0
faker==22.6.0

# Wearable Integrations
fitbit==0.3.1
python-apple-health==0.1.0

# Utilities
pydantic==2.6.0
python-dateutil==2.8.2
pytz==2024.1
schedule==1.2.1

# NLP
nltk==3.8.1
spacy==3.7.2
textblob==0.18.0

# Cloud & Deployment
gunicorn==21.2.0
uvicorn==0.27.0

# Data Processing
scipy==1.12.0
joblib==1.3.2

# Environment
python-decouple==3.8

# Rate Limiting
slowapi==0.1.9

# Validation
email-validator==2.1.0
phonenumbers==8.13.29

# Caching
redis==5.0.1
cachetools==5.3.2

# Async Support
asyncio==3.4.3
aiofiles==23.2.1

# Time & Scheduling
APScheduler==3.10.4

# Image Processing (for charts)
Pillow==10.2.0

# Progress Bars
tqdm==4.66.1

# Configuration
pyyaml==6.0.1
toml==0.10.2

# HTTP Client
urllib3==2.2.0

# Timezone
tzlocal==5.2
